refactor: core/stress_state を dimension にリネームし state との名前重複を解消 (#54)
