+++ src/manforge/core/__init__.py
@@ -1,5 +1,5 @@
 from manforge.core.state import Implicit, Explicit, State, NTENS, StateResidual, StateUpdate, DlambdaResidual
-from manforge.core.stress_state import (
+from manforge.core.dimension import (
     StressState,
     SOLID_3D,
     PLANE_STRAIN,
