+++ src/manforge/core/dimension.py
