+++ tests/integration/test_augmented_residual.py
@@ -22,7 +22,7 @@ import autograd.numpy as anp
 import manforge  # enables JAX float64
 from manforge.models.af_kinematic import AFKinematic3D, AFKinematicPS
 from manforge.models.j2_isotropic import J2Isotropic3D
-from manforge.core.stress_state import PLANE_STRAIN, PLANE_STRESS
+from manforge.core.dimension import PLANE_STRAIN, PLANE_STRESS
 from manforge.simulation.integrator import PythonIntegrator
 from manforge.verification.fd_check import check_tangent
 from tests.fixtures.implicit_models import (
