+++ tests/integration/test_stress_driver.py
@@ -3,7 +3,7 @@
 import numpy as np
 import pytest
 
-from manforge.core.stress_state import SOLID_3D, UNIAXIAL_1D
+from manforge.core.dimension import SOLID_3D, UNIAXIAL_1D
 from manforge.models.j2_isotropic import J2Isotropic3D, J2Isotropic1D
 from manforge.simulation.driver import StressDriver, StrainDriver
 from manforge.simulation.integrator import PythonNumericalIntegrator, PythonAnalyticalIntegrator
